pyproject.toml
[project]
name = "mailzen"
version = "0.1.0"
description = "Clean email box agent"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.1",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.128.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "redis>=7.1.0",
    "sqlalchemy[asyncio]>=2.0.46",
]

[dependency-groups]
dev = [
    "psycopg2-binary>=2.9.11",
]
